cpl_c: convert unsigned to signed check
Force cast unsigned to signed to do a proper check for negative values

Closes coverity CID 150550-150562

+++ modules/cpl_c/cpl_run.c
@@ -91,7 +91,7 @@ static str cpl_302_reason = str_init("Moved temporarily");
 		_code_ = ntohs( *((unsigned short*)(_p_)) );\
 		_n_ =  ntohs( *((unsigned short*)((_p_)+2)) );\
 		/* make sure it remains positive! */\
-		if (_n_ < 0) {\
+		if (((short)(_n_)) < 0) {\
 			LM_ERR("negative attribute length %d\n", _n_);\
 			goto _error_; \
 		} \
